Fix profile-editing bug: input.numeric now gets handled by application.js, when it didn't before, so consolidate and improve the JS for it, and add a validation on Users as well

+++ app/controllers/users_controller.rb
@@ -69,13 +69,9 @@ class UsersController < ApplicationController
         redirect_to '/courses', notice: "Profile successfully updated"
       end
     else
-      FileUtils.rm(up[:profile]) # cleanup orphaned file
-      if current_user_site_admin?
-        render action: "edit"
-      else
-        redirect_to '/main/auth',
-          alert: "Error updating name: #{@user.errors.full_messages.join('; ')}"
-      end
+      FileUtils.rm(up[:profile]) if up[:profile] # cleanup orphaned file
+      render action: "edit", 
+             alert: "Error updating user: #{@user.errors.full_messages.join('; ')}"
     end
   end
